+++ packages/manager/apps/cloud/client/components/at-internet/at-internet.constants.js
@@ -1,17 +1,21 @@
-angular.module('managerApp').constant('TRACKING', {
+export const TRACKING = {
   EU: {
     config: {
-      level2: '10', // 1 is Cloud project ID in AT-Internet EU manager
+      level2: '85',
     },
   },
   CA: {
     config: {
-      level2: '10', // 1 is Cloud project ID in AT-Internet CA manager
+      level2: '85',
     },
   },
   US: {
     config: {
-      level2: '56', // 56 is Cloud project ID in AT-Internet US manager
+      level2: '56',
     },
   },
-});
+};
+
+export default {
+  TRACKING,
+};