+++ packages/manager/apps/dedicated/client/app/dedicated/server/monitoring/monitoring.controller.js
@@ -29,7 +29,9 @@ export default class Monitoring {
       smsNotifications: false,
     };
 
-    this.$scope.$on('server.monitoring.reload', this.refreshTableMonitoring);
+    this.$scope.$on('server.monitoring.reload', () =>
+      this.refreshTableMonitoring(),
+    );
 
     return this.refreshTableMonitoring();
   }
@@ -67,6 +69,8 @@ export default class Monitoring {
   refreshTableMonitoring() {
     this.loaders.monitorings = true;
     this.monitoringServicesIds = [];
+    this.editMode = null;
+    this.addMode = null;
 
     this.Server.getServiceMonitoringIds(this.$stateParams.productId)
       .then((monitoringIds) => {