+++ packages/manager/modules/pci/src/projects/project/project.module.js
@@ -27,7 +27,6 @@ import users from './users';
 import vouchers from './vouchers';
 import regions from './regions';
 import routing from './project.routing';
-import streams from './streams';
 import serving from './serving';
 import training from './training';
 import workflow from './workflow';
@@ -62,7 +61,6 @@ angular
     storages,
     users,
     vouchers,
-    streams,
     workflow,
     serving,
     training,